api/requirements.txt
requests==2.31.0
opencv-python==4.9.0.80
scikit-learn==1.4.1.post1
flask-cors==4.0.0
flask==3.0.2
face-recognition==1.3.0
ultralytics==8.1.34
pillow==10.2.0
numpy==1.26.4